requirements.txt
flask
openai
python-dotenv
wfastcgi
requests
